pass PredicateFilter to compute_bounds
